requirements.txt
confluent-kafka[avro]
pandas
python-dotenv